01 Grundlagen/LED-Laufschrift (Raspberry Pi)/requirements.txt
luma.led_matrix
RPi.GPIO
